测试和控制修改
测试了电机，修改了接口，编码器接口修改。修改了控制模型

+++ Projecct/USER/inc/Global.h
@@ -25,8 +25,8 @@ enum
     White
 };
 
-extern short int XSpeed, YSpeed;
-extern short int XSpeedReal, YSpeedReal;
+extern float XSpeed, YSpeed;
+extern float XSpeedReal, YSpeedReal;
 extern float XPosition, YPosition;
 
 #endif
